backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.6
pydantic>=2.6.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
fugashi>=1.3.0
ipadic>=1.0.0
jamdict>=0.1a11
jamdict-data>=1.5
wn>=0.9.0
wanakana-python>=1.2.0
httpx>=0.26.0
slowapi>=0.1.9
openai>=1.0.0
google-genai>=1.0.0
Pillow>=10.0.0
boto3>=1.34.0
ruff>=0.4.0
